generalize revive button
