+++ packages/client/src/app/components/library/actions/index.ts
@@ -1,4 +1,5 @@
 export { CollectButton } from './CollectButton';
 export { FeedButton } from './FeedButton';
 export { LiquidateButton } from './LiquidateButton';
+export { ReviveButton } from './ReviveButton';
 export { StopButton } from './StopButton';
